.. Hmm this should be in a diff branch I guess I think I have finally tracked down the source of my "randomness" this current example code turns the light on when either sensor reaches the FF error state the problem is that it sometimes reaches this as we expected for long range, but also sometimes for short range.
this is a problem and I think it has to do with the 100 microseconds we are throwing away at the begginning, I am going to try to tune that down untill I get less close up garbage data. but this might have to change on a per sensor base when we get close.